+++ src/app/services/datasource.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Subject } from 'rxjs';
+import { lastValueFrom, Subject, take } from 'rxjs';
 import { NanoClient } from '@dev-ptera/nano-node-rpc';
 import { AppStateService } from '@app/services/app-state.service';
 import { SELECTED_RPC_DATASOURCE_CHANGE } from '@app/services/wallet-events.service';
@@ -45,6 +45,9 @@ export class DatasourceService {
     defaultRpcDataSource: Datasource;
     customRpcDataSources: Datasource[] = [];
 
+    defaultSpyglassSource: Datasource;
+    customSpyglassSources: Datasource[] = [];
+
     private rpcNode: NanoClient;
     private rpcSource: Datasource;
     private readonly rpcSourceLoadedSubject = new Subject<Datasource>();
@@ -52,69 +55,32 @@ export class DatasourceService {
     private spyglassApiSource: Datasource;
     private readonly spyglassSourceLoadedSubject = new Subject<Datasource>();
 
-    constructor(http: HttpClient, state: AppStateService) {
-        const handleError = (err, url): void => {
-            console.error(`${url} is inaccessible as a datasource, ignoring it.`);
-            console.error(err);
-        };
+    handleError = (err, url): void => {
+        console.error(`${url} is inaccessible as a datasource, ignoring it.`);
+        console.error(err);
+    };
 
-        state.store.subscribe((data) => {
-            if (!data.customRpcNodeURLs || data.customRpcNodeURLs.length === this.customRpcDataSources.length) {
-                return;
-            }
+    constructor(private _http: HttpClient, private _state: AppStateService) {}
 
-            this.customRpcDataSources = [];
-            data.customRpcNodeURLs.map((customSourceUrl, index) => {
-                const newSource = {
-                    isSelected: false,
-                    isAccessible: true,
-                    isAddedByUser: true,
-                    alias: `Custom node #${index + 1}`,
-                    url: customSourceUrl,
-                };
-                this.customRpcDataSources.push(newSource);
-            });
-
-            /** When a custom datasource is added, we will set it as the selected RPC datasource.
-             *  There is currently no check in place, within the component, to check to see if the datasource is online when adding it.
-             * */
-            this.setRpcSource(
-                this.customRpcDataSources[this.customRpcDataSources.length - 1] || this.defaultRpcDataSource
-            );
-        });
+    init(): void {
+        // eslint-disable-next-line no-console
+        console.info('Datasource Service Initialized');
+        const store = this._state.store.getValue();
 
         // Ping available RPC Sources
-        this.availableRpcDataSources.map((source: Datasource) => {
-            http.post(source.url, { action: 'block_count' })
-                .toPromise()
-                .then(() => {
-                    source.isAccessible = true;
-                    /** We want to default the RPC node to Kalium since it is configured to work with Boom-PoW. */
-                    if (!this.rpcSource || (source.alias === 'Kalium' && !this._isCustomSource(this.rpcSource))) {
-                        // eslint-disable-next-line no-console
-                        console.log(`Using ${source.alias} as RPC source.`);
-                        this.setRpcSource(source);
-                        this.defaultRpcDataSource = source;
-                        this.rpcSourceLoadedSubject.next(source);
-                    }
-                })
-                .catch((err) => handleError(err, source.url));
+        store.customRpcNodeSources.map((url) => {
+            this.addCustomDatasource('rpc', url, false);
+        });
+        [...this.availableRpcDataSources, ...this.customRpcDataSources].map((source: Datasource) => {
+            this._checkRpcSourceOnline(source);
         });
 
         // Ping available Spyglass API sources
-        this.availableSpyglassApiSources.map((source) => {
-            http.get<any>(`${source.url}/v1/representatives/online`)
-                .toPromise()
-                .then(() => {
-                    source.isAccessible = true;
-                    if (!this.spyglassApiSource) {
-                        // eslint-disable-next-line no-console
-                        console.log(`Using ${source.alias} as Spyglass API source.`);
-                        this.setSpyglassApiSource(source);
-                        this.spyglassSourceLoadedSubject.next(source);
-                    }
-                })
-                .catch((err) => handleError(err, source.url));
+        store.customSpyglassApiSources.map((url) => {
+            this.addCustomDatasource('spyglass', url, false);
+        });
+        [...this.availableSpyglassApiSources, ...this.customSpyglassSources].map((source: Datasource) => {
+            this._checkSpyglassSourceOnline(source);
         });
     }
 
@@ -176,5 +142,97 @@ export class DatasourceService {
                 });
             }
         });
+    }
+
+    private _checkRpcSourceOnline(source: Datasource): void {
+        const req$ = this._http.post(source.url, { action: 'block_count' }).pipe(take(1));
+        lastValueFrom(req$)
+            .then(() => {
+                source.isAccessible = true;
+                /** RPC Priority order:
+                 *  1. Custom sources
+                 *  2. Kalium (enables Boom-PoW)
+                 *  3. Booster
+                 */
+                if (!this._isCustomSource(source)) {
+                    if (!this.defaultRpcDataSource || source.alias === 'Kalium') {
+                        this.defaultRpcDataSource = source; // Default needs to be one that can't be removed.
+                    }
+                }
+
+                if (!this._isCustomSource(source) && this.rpcSource && this._isCustomSource(this.rpcSource)) {
+                    return;
+                }
+
+                if (!this.rpcSource || source.alias === 'Kalium' || this._isCustomSource(source)) {
+                    // eslint-disable-next-line no-console
+                    console.log(`Using ${source.alias} as RPC source.`);
+                    this.setRpcSource(source);
+                    this.rpcSourceLoadedSubject.next(source);
+                }
+            })
+            .catch((err) => this.handleError(err, source.url));
+    }
+
+    private _checkSpyglassSourceOnline(source: Datasource): void {
+        const req$ = this._http.get<any>(`${source.url}/v1/representatives/online`).pipe(take(1));
+        lastValueFrom(req$)
+            .then(() => {
+                source.isAccessible = true;
+                /** Spyglass Priority order:
+                 *  1. Custom sources
+                 *  2. Batman || Creeper
+                 */
+                if (!this._isCustomSource(source)) {
+                    this.defaultSpyglassSource = source; // Default needs to be one that can't be removed.
+                }
+                if (!this.spyglassApiSource || this._isCustomSource(source)) {
+                    // eslint-disable-next-line no-console
+                    console.log(`Using ${source.alias} as Spyglass API source.`);
+                    this.setSpyglassApiSource(source);
+                    this.spyglassSourceLoadedSubject.next(source);
+                }
+            })
+            .catch((err) => this.handleError(err, source.url));
+    }
+
+    addCustomDatasource(type: 'spyglass' | 'rpc', customSourceUrl: string, checkIfOnline = true): void {
+        const alias = `Custom node #${
+            type === 'rpc' ? this.customRpcDataSources.length + 1 : this.customSpyglassSources.length + 1
+        }`;
+        const datasource = {
+            isSelected: false,
+            isAccessible: false,
+            isAddedByUser: true,
+            alias: alias,
+            url: customSourceUrl,
+        };
+        if (type === 'spyglass') {
+            this.customSpyglassSources.push(datasource);
+        } else {
+            this.customRpcDataSources.push(datasource);
+        }
+
+        /** When a custom datasource is added, we will set it as the selected data source. */
+        if (checkIfOnline) {
+            if (type === 'spyglass') {
+                this._checkSpyglassSourceOnline(datasource);
+            } else {
+                this._checkRpcSourceOnline(datasource);
+            }
+        }
+    }
+
+    removeCustomRpcSource(index: number): void {
+        this.customRpcDataSources.splice(index, 1);
+        this.setRpcSource(this.customRpcDataSources[this.customRpcDataSources.length - 1] || this.defaultRpcDataSource);
+    }
+
+    removeCustomSpyglassSource(index: number): void {
+        this.customSpyglassSources.splice(index, 1);
+        debugger;
+        this.setSpyglassApiSource(
+            this.customSpyglassSources[this.customSpyglassSources.length - 1] || this.defaultSpyglassSource
+        );
     }
 }